Adding new agent/auth wrappers and monitor_agents check in monitord
