+++ src/unit_tests/wrappers/wazuh/shared/agent_op_wrappers.h
@@ -1,0 +1,17 @@
+/* Copyright (C) 2015-2020, Wazuh Inc.
+ * All rights reserved.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation
+ */
+
+
+#ifndef AGENT_OP_WRAPPERS_H
+#define AGENT_OP_WRAPPERS_H
+
+int __wrap_auth_connect();
+char* __wrap_get_agent_id_from_name(__attribute__((unused)) char *agent_name);
+
+#endif
